Fix filter method to handle list/hash values
